Done 8

+++ PythonHomeWork5/PythonHomeWork5.py
@@ -11,6 +11,20 @@ from PythonHomeWork3.victory import victory_game
 from PythonHomeWork4.use_functions import money_game
 from PythonHomeWork1.PythonHomeWork1 import about
 
+def add_separators(f):
+    # inner - итоговая функция с новым поведение
+    def inner(*args, **kwargs):
+        # поведение до вызова
+        print('*' * 10)
+        result = f(*args, **kwargs)
+        # поведение после вызова
+        print('*' * 10)
+        return result
+
+    # возвращается функция inner с новым поведением
+    return inner
+
+@add_separators
 def get_dirs():
     dirs = []
     if os.path.exists(FILE_NAME):
@@ -18,6 +32,7 @@ def get_dirs():
             dirs = pickle.load(f)
     return dirs
 
+@add_separators
 def save_dirsfiles(indirs, infiles):
     with open(FILE_NAME, 'w') as f:
         files = "files: "
@@ -28,6 +43,7 @@ def save_dirsfiles(indirs, infiles):
             files += i+", "
         f.write(files)
 
+@add_separators
 def create_dir():
     dir_name = input("Введите имя новой папки: ")
     path = os.path.join(os.getcwd(),dir_name)
@@ -38,6 +54,7 @@ def create_dir():
     else:
         print("Такая папка уже существет")
 
+@add_separators
 def remove_dir():
     dir_name = input("Введите имя удаляемой папки: ")
     path = os.path.join(os.getcwd(),dir_name)
@@ -48,30 +65,39 @@ def remove_dir():
     else:
         print("Такая папка не существет")
 
+@add_separators
 def about_me():
     print("Автор Волков К.А.")
 
+@add_separators
 def watch_all():
     print(os.listdir())
 
+@add_separators
 def watch_files():
     files = []
-    with os.scandir(os.getcwd()) as listOfEntries:  
-        for entry in listOfEntries:        
-            if entry.is_file():
-                print(entry.name)
-                files.append(entry.name)
+    with os.scandir(os.getcwd()) as listOfEntries:
+        files = [entry.name for entry in listOfEntries if entry.is_file()] #генератор
+        print(files)
+        #for entry in listOfEntries:        
+        #    if entry.is_file():
+        #        print(entry.name)
+        #        files.append(entry.name)
     return files
 
+@add_separators
 def watch_dirs():
      dirs = []
      with os.scandir(os.getcwd()) as listOfEntries:  
-        for entry in listOfEntries:        
-            if not entry.is_file():
-                print(entry.name)
-                dirs.append(entry.name)
+        dirs = [entry.name for entry in listOfEntries if not entry.is_file()]  #генератор
+        print(dirs)
+        #for entry in listOfEntries:        
+        #    if not entry.is_file():
+        #        print(entry.name)
+        #        dirs.append(entry.name)
      return dirs
 
+@add_separators
 def copy_file():
     name = input("Введите имя файла/папки: для копирования ")
     path = os.path.join(os.getcwd(),name)
@@ -80,11 +106,13 @@ def copy_file():
         return
     name_cpy = input("Введите новое имя файла/папки: для копирования ")
     path = os.path.join(os.getcwd(),name_cpy)
-    if not os.path.exists(path):
-        shutil.copy(name, name_cpy)
-    else:
-        print("Такой файл/папка уже существует")
+    shutil.copy(name, name_cpy) if not os.path.exists(path) else print("Такой файл/папка уже существует")  #тернарный оператор
+    #if not os.path.exists(path):
+    #    shutil.copy(name, name_cpy)
+    #else:
+    #    print("Такой файл/папка уже существует")
 
+@add_separators
 def change_dir():
     name = input("Введите имя новой директории ")
     if os.path.exists(name):
